Выделение структур с моделями ячеек в отдельные файлы
