+++ FakeNFT/Scenes /Catalog/NftCatalog/Cell/CatalogNftCellModel.swift
@@ -1,0 +1,14 @@
+//
+//  CatalogNftCellModel.swift
+//  FakeNFT
+//
+//  Created by Эмилия on 01.03.2024.
+//
+
+import Foundation
+
+struct CatalogNftCellModel {
+    let nameNft: String
+    let countNft: Int
+    let url: URL
+}
