+++ FakeNFT/Scenes /Catalog/NftCollection/Cell/NftCollectionCellModel.swift
@@ -1,0 +1,18 @@
+//
+//  NftCollectionCellModel.swift
+//  FakeNFT
+//
+//  Created by Эмилия on 01.03.2024.
+//
+
+import Foundation
+
+struct NftCollectionCellModel {
+    let id: String
+    let nameNft: String
+    let price: Float
+    let isLiked: Bool
+    let isInTheCart: Bool
+    let rating: Int
+    let url: URL
+}
